Se arreglo la funcionalidad para agregar asignaciones a usuarios con menos tareas en curso

+++ reportes/delete_report.php
@@ -1,10 +1,17 @@
 <?php
 include "../db/db_connection.php";
 $id = $_GET['id'];
-$sql = "DELETE FROM `reporte` WHERE id_reporte = $id";
-$result = mysqli_query($connection, $sql);
-if ($result) {
-    header("Location: reportes.php?msg=Laboratorio eliminado");
+$sql_asignacion = "DELETE FROM `asignacion` WHERE id_reporte = $id";
+$result_asignacion = mysqli_query($connection, $sql_asignacion);
+if ($result_asignacion) {
+    $sql = "DELETE FROM `reporte` WHERE id_reporte = $id";
+    $result = mysqli_query($connection, $sql);
+    if ($result) {
+        header("Location: reportes.php?msg=Reporte eliminado");
+    } else {
+        echo "Error: " . mysqli_error($connection);
+    }
 } else {
     echo "Error: " . mysqli_error($connection);
 }
+?>
